Better clean up in tests

+++ plugins/woocommerce/tests/php/includes/class-wc-tax-test.php
@@ -7,6 +7,8 @@
 
 declare( strict_types=1 );
 
+use ReflectionClass;
+
 /**
  * Unit tests for the WC_Tax class get_shipping_tax_rates method and related functionality.
  * Covers edge case bug from https://github.com/woocommerce/woocommerce/issues/58757
@@ -35,6 +37,13 @@ class WC_Tax_Test extends WC_Unit_Test_Case {
 	private $created_products = array();
 
 	/**
+	 * Store the original value of woocommerce_calc_taxes.
+	 *
+	 * @var string|null
+	 */
+	private $original_calc_taxes = null;
+
+	/**
 	 * Set up test.
 	 */
 	public function setUp(): void {
@@ -43,12 +52,8 @@ class WC_Tax_Test extends WC_Unit_Test_Case {
 		// Clear any existing tax rates and settings.
 		$this->clean_up_tax_rates();
 
-		// Set up base store location.
-		update_option( 'woocommerce_default_country', 'US:CA' );
-		update_option( 'woocommerce_store_address', '123 Test St' );
-		update_option( 'woocommerce_store_postcode', '90210' );
-
-		// Enable tax calculations.
+		// Store and set woocommerce_calc_taxes option.
+		$this->original_calc_taxes = get_option( 'woocommerce_calc_taxes', null );
 		update_option( 'woocommerce_calc_taxes', 'yes' );
 
 		// Ensure required tax classes exist for testing.
@@ -79,7 +84,13 @@ class WC_Tax_Test extends WC_Unit_Test_Case {
 		// Reset tax settings.
 		delete_option( 'woocommerce_shipping_tax_class' );
 		delete_option( 'woocommerce_tax_classes' );
-		delete_option( 'woocommerce_calc_taxes' );
+
+		// Restore woocommerce_calc_taxes option to its original value.
+		if ( null !== $this->original_calc_taxes ) {
+			update_option( 'woocommerce_calc_taxes', $this->original_calc_taxes );
+		} else {
+			delete_option( 'woocommerce_calc_taxes' );
+		}
 
 		// Only clean up tax classes we created during testing.
 		if ( $this->created_reduced_rate_class ) {
@@ -98,6 +109,7 @@ class WC_Tax_Test extends WC_Unit_Test_Case {
 		$wpdb->query( "DELETE FROM {$wpdb->prefix}woocommerce_tax_rate_locations" );
 		$wpdb->query( "DELETE FROM {$wpdb->prefix}woocommerce_tax_rates" );
 		wp_cache_flush();
+		\WC_Cache_Helper::invalidate_cache_group( 'taxes' );
 	}
 
 	/**
@@ -149,10 +161,10 @@ class WC_Tax_Test extends WC_Unit_Test_Case {
 	public function test_get_shipping_tax_rates_with_standard_tax_class() {
 		// Create a standard tax rate with shipping enabled.
 		$tax_rate = array(
-			'tax_rate_country'  => 'US',
-			'tax_rate_state'    => 'CA',
+			'tax_rate_country'  => '',
+			'tax_rate_state'    => '',
 			'tax_rate'          => '10.0000',
-			'tax_rate_name'     => 'CA Sales Tax',
+			'tax_rate_name'     => 'Test Sales Tax',
 			'tax_rate_priority' => '1',
 			'tax_rate_compound' => '0',
 			'tax_rate_shipping' => '1',
@@ -173,7 +185,7 @@ class WC_Tax_Test extends WC_Unit_Test_Case {
 
 		$this->assertArrayHasKey( $tax_rate_id, $shipping_rates );
 		$this->assertEquals( '10.0000', $shipping_rates[ $tax_rate_id ]['rate'] );
-		$this->assertEquals( 'CA Sales Tax', $shipping_rates[ $tax_rate_id ]['label'] );
+		$this->assertEquals( 'Test Sales Tax', $shipping_rates[ $tax_rate_id ]['label'] );
 		$this->assertEquals( 'yes', $shipping_rates[ $tax_rate_id ]['shipping'] );
 
 		WC_Tax::_delete_tax_rate( $tax_rate_id );
@@ -185,8 +197,8 @@ class WC_Tax_Test extends WC_Unit_Test_Case {
 	public function test_get_shipping_tax_rates_with_reduced_tax_class() {
 		// Create reduced rate tax with shipping enabled.
 		$reduced_tax_rate = array(
-			'tax_rate_country'  => 'US',
-			'tax_rate_state'    => 'CA',
+			'tax_rate_country'  => '',
+			'tax_rate_state'    => '',
 			'tax_rate'          => '5.0000',
 			'tax_rate_name'     => 'Reduced Tax',
 			'tax_rate_priority' => '1',
@@ -222,8 +234,8 @@ class WC_Tax_Test extends WC_Unit_Test_Case {
 	public function test_get_shipping_tax_rates_edge_case_no_reduced_shipping_rates() {
 		// Create standard tax rate with shipping enabled.
 		$standard_tax_rate = array(
-			'tax_rate_country'  => 'US',
-			'tax_rate_state'    => 'CA',
+			'tax_rate_country'  => '',
+			'tax_rate_state'    => '',
 			'tax_rate'          => '10.0000',
 			'tax_rate_name'     => 'Standard Tax',
 			'tax_rate_priority' => '1',
@@ -235,8 +247,8 @@ class WC_Tax_Test extends WC_Unit_Test_Case {
 
 		// Create reduced rate tax WITHOUT shipping enabled (shipping = 0).
 		$reduced_tax_rate = array(
-			'tax_rate_country'  => 'US',
-			'tax_rate_state'    => 'CA',
+			'tax_rate_country'  => '',
+			'tax_rate_state'    => '',
 			'tax_rate'          => '5.0000',
 			'tax_rate_name'     => 'Reduced Tax',
 			'tax_rate_priority' => '1',
@@ -272,8 +284,8 @@ class WC_Tax_Test extends WC_Unit_Test_Case {
 	public function test_get_shipping_tax_rates_with_explicit_shipping_tax_class() {
 		// Create reduced rate tax with shipping enabled.
 		$reduced_tax_rate = array(
-			'tax_rate_country'  => 'US',
-			'tax_rate_state'    => 'CA',
+			'tax_rate_country'  => '',
+			'tax_rate_state'    => '',
 			'tax_rate'          => '5.0000',
 			'tax_rate_name'     => 'Reduced Tax',
 			'tax_rate_priority' => '1',
@@ -329,8 +341,8 @@ class WC_Tax_Test extends WC_Unit_Test_Case {
 	public function test_get_shipping_tax_rates_mixed_tax_classes_prioritizes_standard() {
 		// Create both standard and reduced rate taxes with shipping enabled.
 		$standard_tax_rate = array(
-			'tax_rate_country'  => 'US',
-			'tax_rate_state'    => 'CA',
+			'tax_rate_country'  => '',
+			'tax_rate_state'    => '',
 			'tax_rate'          => '10.0000',
 			'tax_rate_name'     => 'Standard Tax',
 			'tax_rate_priority' => '1',
@@ -341,8 +353,8 @@ class WC_Tax_Test extends WC_Unit_Test_Case {
 		);
 
 		$reduced_tax_rate = array(
-			'tax_rate_country'  => 'US',
-			'tax_rate_state'    => 'CA',
+			'tax_rate_country'  => '',
+			'tax_rate_state'    => '',
 			'tax_rate'          => '5.0000',
 			'tax_rate_name'     => 'Reduced Tax',
 			'tax_rate_priority' => '1',
